Add files via upload
This is a Google Collab notebook.  Upload it there then go to Edit->Notebook Settings  and enable the GPU to speed up the code.

Most of the code to show the different tests done in commented out down there.  Showing normal and multithreaded attempts for building the neural network and trying to have it identify images.  It also has some results related code, however, that wasn't really tested since the multithreaded stuff didn't work, so it never got to the point where that code was run.  There is also the function to run the multi_gpu_model version of creating the neural network, but it won't work unless you move this to your own computer and you have 2 GPUs.  That code will never work on Collab.